refactor AstGraphVariableRenamer with samplepass
